String-ify `replace-vars` args
Prior to this commit, trying to insert a number via `replace-vars`
would throw an exception, because the replacement value is not a
string.  Example:

    ;; html
    <div>${a}</div>

    ;; clj
    [:div text-node] (replace-vars {:a 42})

    ;; throws a ClassCastException in clojure.string/replace-by

This commit makes it so that any substitute is converted into its
string representation via `str` before substitution.  For example, 42
is converted into "42" so it can be substituted properly.  This
fixes the issue and allows any `str`-able value to substitute in.